[core] combine CommitMessage and CommitMessageImpl
