+++ paimon-core/src/test/java/org/apache/paimon/table/sink/CommitMessageSerializerTest.java
@@ -43,11 +43,9 @@ public class CommitMessageSerializerTest {
         CompactIncrement compactIncrement = randomCompactIncrement();
         IndexIncrement indexIncrement =
                 new IndexIncrement(Arrays.asList(randomIndexFile(), randomIndexFile()));
-        CommitMessageImpl committable =
-                new CommitMessageImpl(
-                        row(0), 1, newFilesIncrement, compactIncrement, indexIncrement);
-        CommitMessageImpl newCommittable =
-                (CommitMessageImpl) serializer.deserialize(2, serializer.serialize(committable));
+        CommitMessage committable =
+                new CommitMessage(row(0), 1, newFilesIncrement, compactIncrement, indexIncrement);
+        CommitMessage newCommittable = serializer.deserialize(2, serializer.serialize(committable));
         assertThat(newCommittable.compactIncrement()).isEqualTo(committable.compactIncrement());
         assertThat(newCommittable.newFilesIncrement()).isEqualTo(committable.newFilesIncrement());
         assertThat(newCommittable.indexIncrement()).isEqualTo(committable.indexIncrement());
